+++ src/interpreters/__tests__/data/machine2.cov2.test.ts
@@ -1,9 +1,9 @@
 import { castings } from '@bemedev/types';
 import equal from 'fast-deep-equal';
-import { fakeWaiter } from 'src/interpreters/__tests__/fixtures';
 import { _machine2, DELAY, fakeDB } from '~fixturesData';
 import { interpret } from '~interpreters';
 import { nothing } from '~utils';
+import { fakeWaiter } from '../fixtures';
 
 describe('machine coverage', () => {
   beforeAll(() => {